may have a fractional component in the future

+++ v2/forecast.go
@@ -48,7 +48,7 @@ type DataPoint struct {
 	ApparentTemperatureMax     float64 `json:"apparentTemperatureMax,omitempty"`
 	ApparentTemperatureMaxTime int64   `json:"apparentTemperatureMaxTime,omitempty"`
 	NearestStormBearing        float64 `json:"nearestStormBearing,omitempty"`
-	NearestStormDistance       int64   `json:"nearestStormDistance,omitempty"`
+	NearestStormDistance       float64 `json:"nearestStormDistance,omitempty"`
 	DewPoint                   float64 `json:"dewPoint,omitempty"`
 	WindSpeed                  float64 `json:"windSpeed,omitempty"`
 	WindBearing                float64 `json:"windBearing,omitempty"`
